refactor: isolate built-in identifier loader
