+++ src/plugin/src/project-index/constants.js
@@ -1,4 +1,8 @@
 export const PROJECT_MANIFEST_EXTENSION = ".yyp";
+export const PROJECT_ROOT_DISCOVERY_ABORT_MESSAGE =
+    "Project root discovery was aborted.";
+export const PROJECT_INDEX_BUILD_ABORT_MESSAGE =
+    "Project index build was aborted.";
 
 const PROJECT_MANIFEST_EXTENSION_LOWER =
     PROJECT_MANIFEST_EXTENSION.toLowerCase();
